o modified node and relationship hashCode implementations, they   now just return the id o fixed some int multiplications in node/relationship/property   store implementations that resulted in values > 2^31 (caused   everything from illegal position to wrong records accessed) o modified persistence window pool not to allocate memory   mapped windows at startup, instead first "refresh bricks" will   allocate (better when we have huge stores). o "refresh bricks" and "expand bricks" implementation in   persistence window pool have been tweaked to perform better   for very large stores
git-svn-id: https://svn.neo4j.org/projects/neo/trunk@241 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ community/src/java/org/neo4j/impl/nioneo/store/LockableWindow.java
@@ -13,7 +13,7 @@ abstract class LockableWindow implements PersistenceWindow
 {
 	public abstract Buffer getBuffer();
 
-	public abstract int position();
+	public abstract long position();
 	
 	public abstract int size();
 	
@@ -30,7 +30,7 @@ abstract class LockableWindow implements PersistenceWindow
 		this.fileChannel = fileChannel;
 	}
 
-	boolean encapsulates( int position )
+	boolean encapsulates( long position )
 	{
 		return position() <= position && position < position() + size();
 	}
